+++ packages/terra-form-textarea/CHANGELOG.md
@@ -8,12 +8,14 @@ Unreleased
 
 ### Changed
 * Updated UpgradeGuide
+* Updated examples to not use `placeholder` prop
 
 ### Removed
 * Removed themeable variable --terra-form-textarea-placeholder-color
 * Removed themeable variable --terra-form-textarea-placeholder-font-style
 * Removed themeable variable --terra-form-textarea-placeholder-disabled-color
 * Removed themeable variable --terra-form-textarea-placeholder-disabled-font-style
+* Removed tests specifically for `placeholder` prop
 
 4.8.0 - (July 7, 2020)
 ------------------